requirements.txt
transformers
torch
Pillow
tf-keras